Refactor get_package_repo_data out of common
This will allow the *_repo.py files to import from common.
