+++ ros_buildfarm/common.py
@@ -20,9 +20,6 @@ try:
 except ImportError:
     from urlparse import urlparse
 
-from .debian_repo import get_debian_repo_index
-from .rpm_repo import get_ros_rpm_repo_index
-
 
 package_format_mapping = {
     'debian': 'deb',
@@ -578,22 +575,6 @@ def get_packages_in_workspaces(workspace_roots, condition_context=None):
         for pkg in pkgs.values():
             pkg.evaluate_conditions(condition_context)
     return pkgs
-
-
-def get_package_repo_data(repository_baseurl, targets, cache_dir):
-    get_index_methods = {
-        'deb': get_debian_repo_index,
-        'rpm': get_ros_rpm_repo_index,
-    }
-
-    data = {}
-    for target in targets:
-        package_format = package_format_mapping[target.os_name]
-        get_index_method = get_index_methods[package_format]
-        index = get_index_method(
-            repository_baseurl, target, cache_dir)
-        data[target] = index
-    return data
 
 
 def get_xunit_publisher_types_and_patterns():
