+++ ros_buildfarm/status_page.py
@@ -27,7 +27,6 @@ import time
 import yaml
 
 from .common import get_os_package_name
-from .common import get_package_repo_data
 from .common import get_release_view_name
 from .common import get_short_arch
 from .common import get_short_os_code_name
@@ -36,6 +35,7 @@ from .common import package_format_mapping
 from .common import Target
 from .config import get_index as get_config_index
 from .config import get_release_build_files
+from .package_repo import get_package_repo_data
 from .status_page_input import get_rosdistro_info
 from .status_page_input import RosPackage
 from .templates import expand_template
